% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_variable.R
\name{get_variable}
\alias{get_variable}
\title{get_variable}
\usage{
get_variable(df, variable = NULL)
}
\arguments{
\item{variable}{name of the variable you want to retrieve from the dataframe}

\item{dataframe}{input metadata dataframe}
}
\value{
dataframe
}
\description{
return a data frame with just the variable of interest. Contains PID and
timepoint and creates a new ID based on these data
}
\examples{

No examples yet
}
