% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_variables.R
\name{get_variables}
\alias{get_variables}
\title{get_variables}
\usage{
get_variables(df, variables = NULL, format = NULL)
}
\arguments{
\item{variables}{a vector of variable names that you want to
retrieve from the dataframe}

\item{format}{the format of the returned dataframe. Long or wide.}

\item{dataframe}{input metadata dataframe}
}
\value{
dataframe
}
\description{
return a data frame with just the variable of interest. Contains PID and
timepoint and creates a new ID based on these data
}
\examples{

No examples yet
}
