% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_data.R
\name{recode_data}
\alias{recode_data}
\title{recode_data}
\usage{
recode_data(dat)
}
\arguments{
\item{dataframe}{dataframe usually read from the exported RedCap database}
}
\value{
dataframe
}
\description{
recodes "Yes" = 1 and "No" = 0 "Don't know" = NA
}
\examples{
# create dummy dataframe
dat <- data.frame(x = letters[1:10],
                  y = sample(c("Yes", "No"), 10, replace = TRUE))

recode_data(dat)
}
